// src/app/core/models/catalogs/users-access-areas-model.ts
export interface IUserAccessAreas {
  usuario: string; //
  nombre: string; //
  rfc?: string;
  curp?: string;
  calle?: string;
  no_interior?: string;
  colonia?: string;
  codigo_postal?: string;
  telefono?: string;
  profesion?: string;
  cve_cargo?: string;
  fec_ingreso_primera_vez?: string;
  dias_vigencia_password?: string;
  fec_ultimo_cambio_password?: string;
  actualizacion_password?: string;
  no_registro: string; //
  email?: string;
  usuario_sirsae?: string;
  envia_correo?: string;
  asigna_atributos?: string;
  clkdet_sirsae?: string;
  exchange_alias?: string;
  clkdet?: string;
  clkid?: string;
  cve_perfil_mim?: string;
  name_ad?: string;
  cve_cargo_ant?: string;
  no_delegacion?: string;
  no_subdelegacion?: string;
  no_departamento?: string;
  asignado?: string;
  no_deleg1?: string;
  no_depart1?: string;
  ultimo_activo?: string;
  descripcion: string; //
  cve_zona_contrato?: string;
  dif_hora?: string;
  etapa_edo?: string;
  cve_zona_vigilancia?: string;
}